latest

+++ src/QueuePerThreadQueuesPublisher.rs
@@ -23,7 +23,7 @@ impl<MessageHandlerArguments: Debug + Copy, E: Debug> QueuePerThreadQueuesPublis
 {
 	/// Allocate a new instance.
 	#[inline(always)]
-	pub fn allocate(hyper_threads: &BitSet<HyperThread>, defaults: &DefaultPageSizeAndHugePageSizes, queue_size_in_bytes: NonZeroU64, inclusive_maximum_bytes_wasted: usize) -> Self
+	pub fn allocate(hyper_threads: &BitSet<HyperThread>, defaults: &DefaultPageSizeAndHugePageSizes, queue_size_in_bytes: NonZeroU64, inclusive_maximum_bytes_wasted: u64) -> Self
 	{
 		Self
 		{
